Lambda/agent_requirements.txt
boto3==1.35.36
botocore==1.35.36
langchain==0.3.7
langchain-aws==0.2.6
langchain-community==0.3.7
faiss-cpu==1.9.0
Pillow==11.0.0
PyJWT==2.8.0
requests==2.31.0
cryptography==41.0.7
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
rapidfuzz==3.5.2